Add some more game logic
It's getting a bit convoluted, and I'm not sure I'm happpy with it.  I
don't think it's composed very well. I've got some other stuff coming in
another patch set that started much less ambitiously, and I'm much
happier with it. There's some ideas that I think are interesting here,
like the error handling. I'm not 100% sold on that, either, but I think
it's worth discussing.