Add game logic for handling army deployments
I mentioned in the previous commit that I was unhappy with how it was
turning out. I opened up a fresh project and started approaching it from
the bottom-up instead of top-down. I feel better about the new stuff. It
seems to do more with less. There are some things I like about the
earlier code, but I think it will be easier to make them meet in the
middle.